Initial commit
This is the first commit, done after I have done a lot of changes and
have done some test prints that have not worked right.

Things that are good with the design so far:
* Size of the pipe hole (even without the tolerance - that was added to
  allow piecing the halves together to work)
* Spacing and size of the holes for the screws
* Dimensions of the block itself (WxH) are good.

Things that still need fixing:

The main thing is that I still need to come up with a good way to split
this in two, and have it join together, fitting the pieces to interlock
in the tight space offered. This has proven challenging, as there is not
enough space to stack 2 thicknesses of the plate.
The current dovetailing approach will not work and should be altered or
removed.

Some ideas:
* Keep the dovetail, but only dovetail the bottom 4-8mm, so there only
  needs to be 4-8mm of overlap before joining them.
* Instead of dovetail, just use a simpler slot like a square peg slot
  with tight tolerance.